Varias alterações no projeto e o Pedido quase completo, e Tem vazamento de memória

+++ aplicacao/uITensPedidoListaHash.pas
@@ -6,7 +6,7 @@ uses
   System.Generics.Collections, uIntensPedidoDTO;
 
 type
-  TItensPedidoListaHash = TObjectDictionary<Integer ,TItensPedidoDTO>;
+  TItensPedidoListaHash = TObjectDictionary<Double ,TItensPedidoDTO>;
 
 implementation
